added friendly delete Modal and SweetAlert Message

+++ resources/js/components/Departments.vue
@@ -35,14 +35,14 @@
                         </table>
                     </div>
 
-                       <!-- Pagination -->
-                       <div class="d-flex justify-content-center" v-if="departmentLinks.length > 3">
+                    <div class="d-flex justify-content-center" v-if="departmentLinks.length > 3">
                         <nav aria-label="Page navigation example">
                             <ul class="pagination">
-                                <li :class="`page-item ${link.active ? 'active':''} ${!link.url ? 'disabled':''}`" v-for="(link,index) in departmentLinks" :key="index"><a class="page-link" href="#" v-html="link.label" @click.prevent="getResults(link)"></a></li>
+                                <li :class="`page-item ${link.active ? 'active' : ''} ${!link.url ? 'disabled' : ''}`" v-for="(link, index) in departmentLinks" :key="index"><a class="page-link" href="#" v-html="link.label" @click.prevent="getResults(link)"></a></li>
                             </ul>
                         </nav>
                     </div>
+
 
                     <!-- Modal -->
                     <div class="modal fade" id="exampleModal" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
@@ -99,11 +99,11 @@
             }
         },
         methods: {
-            getResults(link){
-                if(!link.url || link.active){
+            getResults(link) {
+                if(!link.url || link.active) {
                     return;
                 }else{
-                    this.$store.dispatch('getDepartmentsResults',link)
+                    this.$store.dispatch('getDepartmentsResults', link);
                 }
             },
             createDepartment() {
@@ -134,7 +134,20 @@
                 // }
             },
             deleteDepartment(department) {
-                this.$store.dispatch('deleteDepartment', department)
+                Swal.fire({
+                    title: 'Are you sure?',
+                    text: "You won't be able to revert this!",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Yes, delete it!'
+                    }).then((result) => {
+                        if (result.isConfirmed) {
+                            this.$store.dispatch('deleteDepartment', department)
+                        }
+                })
+
             },
             // testAction() {
             //     this.$store.dispatch('testAction')
@@ -143,7 +156,7 @@
         mounted() {
             this.$store.dispatch('getDepartments')
             this.$store.dispatch('getAuthRolesAndPermissions')
-        },
+        }, 
         computed: {
             departmentLinks() {
                 return this.$store.getters.departmentLinks
@@ -157,6 +170,6 @@
             current_permissions() {
                 return this.$store.getters.current_permissions
             }
-        }
+        } 
     }
 </script>
